Resolve services from ConfigureService method
We always have customer asking how to resolve a service in ConfigureSerivice method or related question. Our docs currently does not covers this part, only a section of how to resolve service in main. Sometimes these question will be redirected to this stackoverflow post https://stackoverflow.com/questions/31863981/how-to-resolve-instance-inside-configureservices-in-asp-net-core .  But the sample code in that thread misses a using and is binding to the context of the original post, so I did some optimize to the sample code and hope this can be added to our documents